refactor: Move reproject_raster_inputs to raster
